libirc.py: retry and ping after read timed out
Currently, if a blocking read is executed without any responses
for a while, a socket.timeout will be triggered and the program
will quit.

This logic is problematic, especially for forwarding bots: the
server will only PING the client if there's no activity from the
client, but since the bot keeps talking, the server has no reason
to PING the client, and a read timeout will be triggered due to
lack of PING even if the connection is normal.

As a workaround, we put the receive logic in a big loop that runs
only once, but if a socket.timeout is encountered, the program will
PING itself, and reset the loop to the start, and a counter "retry"
is increased by one. The "retry" counter is cleared if we received
data.

If socket.timeout is encountered again and we have tried too many times,
the problem will quit, like the original behavior.

Signed-off-by: Tom Li <biergaizi@member.fsf.org>